requirements.txt
numpy
scipy
matplotlib
joblib
sobol_seq
yfinance
alpha_vantage
torch
gpytorch
PyYAML
scikit-learn
plotly_express
seaborn
arch
scikit-optimize
torchdiffeq
tensorly
mlflow
giotto-tda
scikit-tda
skorch
gudhi
shap
numba